create cvi projection extension
