+++ Project.toml
@@ -37,10 +37,12 @@ Unrolled = "9602ed7d-8fef-5bc8-8597-8f21381861e8"
 [weakdeps]
 Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
+ExponentialFamilyProjection = "17f509fa-9a96-44ba-99b2-1c5f01f0931b"
 
 [extensions]
 ReactiveMPOptimisersExt = "Optimisers"
 ReactiveMPRequiresExt = "Requires"
+ReactiveMPProjectionExt = "ExponentialFamilyProjection"
 
 [compat]
 BayesBase = "1.1.0"
@@ -50,6 +52,7 @@ Distributions = "0.24, 0.25"
 DomainIntegrals = "0.3.2, 0.4"
 DomainSets = "0.5.2, 0.6, 0.7"
 ExponentialFamily = "1.2.0"
+ExponentialFamilyProjection = "1.1"
 FastCholesky = "1.3.0"
 FastGaussQuadrature = "0.4, 0.5"
 FixedArguments = "0.1"
