+++ ext/ReactiveMPProjectionExt/rules/marginals.jl
