--- src/approximations/cvi_projection.jl
@@ -1,3 +1,0 @@
-export CVIProjection
-
-struct CVIProjection <: AbstractApproximationMethod end
