+++ src/rules/predefined.jl
@@ -179,8 +179,4 @@ include("delta/cvi/in.jl")
 include("delta/cvi/out.jl")
 include("delta/cvi/marginals.jl")
 
-include("delta/cvi_projection/in.jl")
-include("delta/cvi_projection/out.jl")
-include("delta/cvi_projection/marginals.jl")
-
 include("half_normal/out.jl")
